Reduces test vectors down to just ethers and passes them all.
Sign, decode, and encode all tested for legacy, 155, 2930, 1559 transactions.

I removed the other test vectors because they were redundant and harder to work with, plus ethers vectors is more than enough for the needs of this library at this time.

There is a bug in the test vectors that results in many not being able to run against EIP-155, hopefully this will get fixed and those vectors will get updated.